Deploy sprint 138

+++ FBOLinx.Web/FBOLinx.Web/Controllers/CustomCustomerTypesController.cs
@@ -182,39 +182,50 @@ namespace FBOLinx.Web.Controllers
         [HttpPost("updatedefaulttemplate")]
         public async Task<ActionResult<PricingTemplate>> UpdateCustomCustomerTypeCollection(UpdateTemplateVM updateTemplate)
         {
-           if(updateTemplate != null)
+           if (updateTemplate != null)
             {
                 var currentDefault = await _context.PricingTemplate.FirstOrDefaultAsync(s => s.Fboid == updateTemplate.fboid && s.Oid == updateTemplate.currenttemplate);
 
                 var newDefault = await _context.PricingTemplate.FirstOrDefaultAsync(s => s.Fboid == updateTemplate.fboid && s.Oid == updateTemplate.newtemplate);
 
-                if(newDefault != null)
+                if (newDefault != null)
                 {
                     newDefault.Default = true;
                 }
 
-                if(currentDefault != null)
+                if (currentDefault != null && currentDefault.Oid != newDefault.Oid)
                 {
-                    _context.PricingTemplate.Remove(currentDefault);
-                }
+                    if (updateTemplate.isDeleting)
+                    {
+                        _context.PricingTemplate.Remove(currentDefault);
+                    }
+                    else
+                    {
+                        currentDefault.Default = false;
+                        _context.PricingTemplate.Update(currentDefault);
+                    }
 
-                if (newDefault != null && currentDefault != null)
-                {
-                    var customers = await _context.CustomCustomerTypes
-                    .Where(c => c.Fboid.Equals(updateTemplate.fboid) && c.CustomerType.Equals(updateTemplate.currenttemplate))
-                    .Select(s => s.CustomerId)
-                    .ToListAsync();
+                    if (updateTemplate.isDeleting)
+                    {
+                        if (newDefault != null && currentDefault != null)
+                        {
+                            var customers = await _context.CustomCustomerTypes
+                            .Where(c => c.Fboid.Equals(updateTemplate.fboid) && c.CustomerType.Equals(updateTemplate.currenttemplate))
+                            .Select(s => s.CustomerId)
+                            .ToListAsync();
 
-                    var groupInfo = await _context.Fbos.FirstOrDefaultAsync(s => s.Oid == updateTemplate.fboid);
-                    var customerInfoByGroup = await _context.CustomerInfoByGroup.Where(s => customers.Contains(s.CustomerId) && s.GroupId == groupInfo.GroupId).ToListAsync();
+                            var groupInfo = await _context.Fbos.FirstOrDefaultAsync(s => s.Oid == updateTemplate.fboid);
+                            var customerInfoByGroup = await _context.CustomerInfoByGroup.Where(s => customers.Contains(s.CustomerId) && s.GroupId == groupInfo.GroupId).ToListAsync();
 
-                    customerInfoByGroup.ForEach(s => s.PricingTemplateRemoved = true);
+                            customerInfoByGroup.ForEach(s => s.PricingTemplateRemoved = true);
 
-                    var customCustomerTypes = await _context.CustomCustomerTypes
-                        .Where(c => c.Fboid.Equals(updateTemplate.fboid) && c.CustomerType.Equals(updateTemplate.currenttemplate))
-                        .ToListAsync();
+                            var customCustomerTypes = await _context.CustomCustomerTypes
+                                .Where(c => c.Fboid.Equals(updateTemplate.fboid) && c.CustomerType.Equals(updateTemplate.currenttemplate))
+                                .ToListAsync();
 
-                    customCustomerTypes.ForEach(c => c.CustomerType = newDefault.Oid);
+                            customCustomerTypes.ForEach(c => c.CustomerType = newDefault.Oid);
+                        }
+                    }
                 }
                 
                 _context.PricingTemplate.Update(newDefault);
